+++ PlataSistem/PlataApp/Views/Dashboard/DashboardPage.xaml.cs
@@ -1,0 +1,11 @@
+using System.Windows.Controls;
+
+namespace PlataApp.Views.Dashboard;
+
+public partial class DashboardPage : Page
+{
+    public DashboardPage()
+    {
+        InitializeComponent();
+    }
+}